% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_extract_topic.R
\name{po_extract_topic}
\alias{po_extract_topic}
\title{create a .pot file out from a single Rd help topic}
\usage{
po_extract_topic(topic, package)
}
\arguments{
\item{topic}{the .Rd help topic}

\item{package}{the package to find the Rd file in}
}
\description{
Grabs the \code{topic.Rd} file and creates a \verb{/po/topic.pot} file ready to translate.
}
\examples{
 po_extract_topic("verbatim_logical", "yaml")

}
