% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rd_flat}
\alias{print.Rd_flat}
\title{Pretty printing for flattened Rd files}
\usage{
\method{print}{Rd_flat}(x, ...)
}
\arguments{
\item{x}{an flattened rd object.}

\item{...}{ignored.}
}
\description{
Pretty printing for flattened Rd files
}
