% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-translate.R
\name{rd_flat_read}
\alias{rd_flat_read}
\title{Read a translated file}
\usage{
rd_flat_read(file)
}
\arguments{
\item{file}{Yaml file with translations}
}
\description{
Reads a yaml file with translations into a list. This list will then be used
to translate the strings at runtime.
}
