% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-flatten.R, R/rd-unflatten.R
\name{rd_flatten}
\alias{rd_flatten}
\alias{rd_unflatten}
\title{Reformats the Rd structure}
\usage{
rd_flatten(
  Rd,
  untranslatable = c("alias", "name", "keyword", "concept", "usage")
)

rd_unflatten(rd_flat)
}
\arguments{
\item{Rd}{a parsed Rd file returned by \code{tools::parse_Rd} or
\code{utils::.getHelpfile()}.}

\item{untranslatable}{character vector of fields that should not be translated.}

\item{rd_flat}{a flattened Rd file.}
}
\description{
The Rd structure returned by \code{tools::parse_Rd} and \code{utils::.getHelpfile()}
can to nested and is not ideal for translation. \code{rd_flatten} "flattens"
that structure into a simpler list and \code{rd_unflatten} goes back to the original
structure.
}
\keyword{internal}
