% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-translate.R
\name{rd_translate}
\alias{rd_translate}
\title{Translates an Rd object}
\usage{
rd_translate(Rd, translation)
}
\arguments{
\item{Rd}{Rd object}

\item{translation}{a flattened rd object returned by \verb{[rd_flatten]} or,
more likely, by \verb{[rd_flat_read]}.}
}
\value{
an Rd object with translated strings.
}
\description{
Takes an object returned by \verb{[tools::parse_Rd]} and \code{utils::.getHelpfile()}
and translates the strings.
}
\keyword{internal}
