% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-lang.R
\name{resolve_lang}
\alias{resolve_lang}
\title{Returns matching languages}
\usage{
resolve_lang(languages, target_language)
}
\arguments{
\item{languages}{a character vector of languages.}

\item{target_language}{a single character string with the desired language.}
}
\value{
a logical vector of the same length as \code{languages} indicating which
language is a match for the target_language
}
\description{
Returns matching languages
}
\details{
This should take into account the ISO hierarchy.
i.e.: if target_language is "es_AR", then "es" is good, but "es_AR" is better.
if target language is "es", then "es_AR" is also good.

Right now it returns \code{TRUE} for exact matches only and if there are no matches,
it returns \code{TRUE} for elements in \code{language} that have the same root language
as \code{target_language}. So if \code{target_language} is \code{"es_AR"}, then \code{"es_UY"},
\code{"es_ES"} and \code{"es"} will all return \code{TRUE}.
}
\keyword{internal}
