% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-create.R
\name{i18n_module_create}
\alias{i18n_module_create}
\title{Creates a translation module}
\usage{
i18n_module_create(
  module_name = NULL,
  language,
  module_path = file.path(".", module_name),
  package_path,
  rstudio_project = TRUE
)
}
\arguments{
\item{module_name}{Name of the translation module. It needs to be a valid
package name. If missing, it will be created as the name of the package.language.}

\item{language}{Language of the translation module}

\item{module_path}{Path where the module will be created.}

\item{package_path}{Path to the package that will be translated.}

\item{rstudio_project}{Logical indicating whether to create an .Rproj file.}
}
\description{
Creates a translation module skeleton form a source package with the
translation templates with the original strings and the proper
functions to load translations.
}
