% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-templates.R
\name{i18n_translation_templates}
\alias{i18n_translation_templates}
\title{Creates templates for translation}
\usage{
i18n_translation_templates(rd_files, po_file)
}
\arguments{
\item{rd_files}{character vector of rd files.}

\item{folder}{folder where to save the templates. Will be created if it
doesn't exist.}
}
\description{
Grabs .Rd files and creates yaml templates to then translate.
}
\examples{
rd_files <- system.file("extdata", "periodic.Rd", package = "rhelpi18n")
po_file <- tempfile()
i18n_translation_templates(rd_files, po_file)

}
