% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-component.R
\name{path_component}
\alias{path_component}
\title{Extract componet of a path}
\usage{
path_component(path, depth = 0)
}
\arguments{
\item{path}{a vector of paths}

\item{depth}{positive integer indicating the component to extract.
0 means the last component, 1 means its parent and so on.}
}
\description{
Extract componet of a path
}
