% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-flat-write.R
\name{rd_flat_write}
\alias{rd_flat_write}
\alias{rd_flat_read}
\title{Write and read a flattened rd file}
\usage{
rd_flat_write(rd_flat, file)

rd_flat_read(file)
}
\arguments{
\item{rd_flat}{an rd_flattened file returned by \verb{[rd_flatten]}.}

\item{file}{Yaml file with translations}
}
\description{
Reads a yaml file with translations into a list. This list will then be used
to translate the strings at runtime.
}
\details{
\code{rd_flat_write} writes the result of \verb{[rd_flatten]} into a yaml file that
then can be used for translation.
\code{rd_flat_read} reads this file and the result can be used as translation
for \verb{[rd_translate]}.
}
